progressbar edit-task
worked on navbar in phonemode => will now hide after click. fixed the title so it doesnt grow and become a big button. change the menu button no longer be <a> as before it added # to the url

card show progress bar and % on how far user has come to achive the goal based on the filter

user can now edit the tracker after creation

change quotation marks marks to be single